finish with /register_payments

+++ api/response/register_payment.py
@@ -5,7 +5,7 @@ from api.base import ResponseDto
 
 class ResponseRegisterPaymentDtoSchema(Schema):
     order_id = fields.Int(required=True)
-    amount = fields.Int(required=True)
+    amount = fields.Float(required=True)
     sberbank_order_id = fields.Str(required=True)
     payment_form_url = fields.Str(required=True)
